+++ com.unity.shadergraph/CHANGELOG.md
@@ -32,6 +32,7 @@ The version number for this package has increased due to a version update of a r
 - Fixed an issue where user setting a property to not Exposed, Hybrid-Instanced would result in a non-Hybrid Global property [1285700]
 - Fixed an issue with Gradient when it is used as expose parameters. Generated code was failing [1285640 ]
 - Fixed the subgraph slot sorting function [1286805]
+- Fixed an issue where dropping HDRP-only properties from the blackboard field into the graph would soft-lock the graph [1288887]
 
 ## [10.1.0] - 2020-10-12